5.06 Breaking two solutions into separate files
